Rename image command to render
