+++ src/usr/mod.rs
@@ -20,7 +20,6 @@ pub mod hex;
 pub mod host;
 pub mod http;
 pub mod httpd;
-pub mod image;
 pub mod install;
 pub mod keyboard;
 pub mod life;
@@ -33,6 +32,7 @@ pub mod pi;
 pub mod pow;
 pub mod r#move;
 pub mod read;
+pub mod render;
 pub mod shell;
 pub mod socket;
 pub mod tcp;
